This refactors the `SearchPageLayoutContext` into multiple components to be more in line with our current approach to managing contexts and hooks.
This was done for implementation consistency across our application.
