+++ graylog2-web-interface/src/views/components/ViewActionsMenu.tsx
@@ -31,7 +31,8 @@ import { ViewStore } from 'views/stores/ViewStore';
 import { SearchMetadataStore } from 'views/stores/SearchMetadataStore';
 import type SearchMetadata from 'views/logic/search/SearchMetadata';
 import * as ViewPermissions from 'views/Permissions';
-import { ViewActionsLayoutOptions, useSearchPageLayout } from 'views/components/contexts/SearchPageLayoutContext';
+import useSearchPageLayout from 'hooks/useSearchPageLayout';
+import { ViewActionsLayoutOptions } from 'views/components/contexts/SearchPageLayoutContext';
 import View from 'views/logic/views/View';
 import type User from 'logic/users/User';
 import CurrentUserContext from 'contexts/CurrentUserContext';
